Added math functionality Added steps for guardians for key ceremony

+++ bindings/netstandard/ElectionGuard/ElectionGuard.ElectionSetup/ElectionKeyPair.cs
@@ -5,7 +5,7 @@
 /// </summary>
 public class ElectionKeyPair
 {
-    public ElectionKeyPair(string guardianId, int sequenceOrder, ElGamalKeyPair keyPair, ElectionPolynomial polynomial)
+    public ElectionKeyPair(string guardianId, ulong sequenceOrder, ElGamalKeyPair keyPair, ElectionPolynomial polynomial)
     {
         OwnerId = guardianId;
         SequenceOrder = sequenceOrder;
@@ -21,7 +21,7 @@ public class ElectionKeyPair
     /// <summary>
     /// The sequence order of the owner guardian
     /// </summary>
-    public int SequenceOrder { get; set; }
+    public ulong SequenceOrder { get; set; }
 
     /// <summary>
     /// The pair of public and private election keys for the guardian
@@ -46,7 +46,7 @@ public class ElectionKeyPair
 
     public static ElectionKeyPair GenerateElectionKeyPair(
         string guardianId,
-        int sequenceOrder,
+        ulong sequenceOrder,
         int quorum,
         ElementModQ nonce)
     {
